requirements.txt
APScheduler==3.10.1
asgiref==3.7.2
Django==4.2.3
django-autoslug==1.9.9
Pillow==10.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==5.0.1
